Changed br argument to int

+++ lib/TricolorLED/TricolorLED.cpp
@@ -103,7 +103,7 @@ int TricolorLED::_rgb_scale(int value) {
   );
 };
 
-void TricolorLED::set_color(int r, int g, int b, float br) {
+void TricolorLED::set_color(int r, int g, int b, int br) {
   /*
     Sets RGB LED colors using 0 - 255 integer scale for each color channel
     and a 0 - 1 float scale for brightness.
